Smart Pixels: Move out of SystemUI
Having the smart pixels service in systemui caused a lot of issues. The
big one being that if systemui was restarted via the AM hack made famous
by Roman, the service can not be restarted by the user unless the devuce
is rebooted. This made for a bad user experience.

Test:
- Enable smart pixels
- Restart systemui (adb shell pkill -TERM -f com.android.systemui)

Change-Id: If374a03876923e112bc3f4f3032e1682f653c1fc
Signed-off-by: Adin Kwok <adin.kwok@carbonrom.org>
Signed-off-by: SahilSonar <sss.sonar2003@gmail.com>